+++ mentor-me-app/src/components/Questions/Button.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+// * STYLE IMPORTS (style-components)
+import style from './StyledQuestions';
+
+// ? BUTTON COMPONENT FOR Questions.js
+export const Button = () => (
+
+    <style.btnDiv>
+        {/* // TODO: IMG icon here.
+        */}
+        
+        <button>Respond</button>
+    </style.btnDiv>
+)
