+++ mentor-me-app/src/components/Questions/Description.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+// * STYLE IMPORTS (style-components)
+import style from './StyledQuestions';
+
+// ? DESCRIPTION COMPONENT FOR Questions.js
+export const Description = () => (
+    <style.descriptionDiv>
+        {/* //* USER POSTS AND DETAILS
+         */}
+        <article className={`description__inner-div`}>
+            <h3>Description</h3>
+            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+        </article>
+    </style.descriptionDiv>
+)
